+++ abstutil/src/random.rs
@@ -5,6 +5,7 @@ use serde_derive::{Deserialize, Serialize};
 
 // Need to explain this trick -- basically keeps consistency between two different simulations when
 // each one might make slightly different sequences of calls to the RNG.
+// TODO Was once used for parked car seeding, but we're not doing things that way anymore...
 pub fn fork_rng(base_rng: &mut XorShiftRng) -> XorShiftRng {
     XorShiftRng::from_seed([base_rng.next_u32() as u8; 16])
 }
