+++ sim/src/make/mod.rs
@@ -1,12 +1,13 @@
 mod a_b_test;
+mod generator;
 mod load;
 mod scenario;
 mod spawner;
 
 pub use self::a_b_test::ABTest;
-pub use self::load::SimFlags;
-pub use self::scenario::{
-    BorderSpawnOverTime, IndividTrip, OriginDestination, PersonSpec, Population, Scenario,
-    SeedParkedCars, SpawnOverTime, SpawnTrip,
+pub use self::generator::{
+    BorderSpawnOverTime, OriginDestination, ScenarioGenerator, SeedParkedCars, SpawnOverTime,
 };
+pub use self::load::SimFlags;
+pub use self::scenario::{IndividTrip, PersonSpec, Scenario, SpawnTrip};
 pub use self::spawner::{TripSpawner, TripSpec};
